chore(cli-go): exclude internal/start from the bundled Go build (CLI-1966)
internal/start (Go's `supabase start`, 1509 LOC) is unreachable from the TS
shell -- native TS start talks to Docker directly, and the only remaining
Go delegation is `db __db-bootstrap` -> internal/db/start, a separate,
small package. Yet it alone drags in an exclusive dependency tree
(docker-compose/v2, buildx, moby/buildkit, k8s client-go, aws-sdk-go-v2)
that measures out to roughly half the shipped supabase-go binary's size
across every release target (darwin/linux/windows, arm64/amd64):
97.6 MB -> 48.7 MB on darwin/arm64 (-48.8 MB, 50.0%); 51-52 MB saved on
every other target.

cmd/start.go keeps the "start" command's registration, flags, and
validation tag-neutral so the shipped command/flag/completion surface can
never drift between builds; only its RunE body differs, split into
cmd/start_full.go (default build, backed by internal/start -- this is
apps/cli-go's own CI and the migration's parity oracle) and
cmd/start_bundled.go (-tags bundled, a stub error instead of importing
internal/start). apps/cli/scripts/build.ts and tools/release/local-release.ts,
the two places that produce the actually-shipped binary, now pass
-tags bundled. cli-go-ci.yml's Start job builds and verifies both variants
(dependency-tree exclusion, command/flag/completion parity, and the stub's
error behavior) before running the existing full e2e against the untagged
tree.

internal/start's source is untouched and keeps building/testing normally
without the tag.
